requirements.txt
aiogram==3.15.0
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
greenlet==3.1.1
pydantic-settings==2.6.1
python-dotenv==1.0.1
redis==5.2.1
alembic==1.14.0
sentry-sdk==2.19.2
gspread==6.1.4
google-auth==2.37.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
